Move test test_ogr2ogr_lib.py
